Extensions for data lists only don't need anymore the _datalist
