+++ src/main/java/net/mcreator/ui/dialogs/SoundElementDialog.java
@@ -60,7 +60,7 @@ public class SoundElementDialog {
 							.collect(Collectors.toList()));
 		}
 
-		JComboBox<String> soundCategory = new JComboBox<>(ElementUtil.loadDataList("soundcategories"));
+		JComboBox<String> soundCategory = new JComboBox<>(ElementUtil.getDataList("soundcategories"));
 
 		ui.add(L10N.label("dialog.sounds.registry_name"));
 		ui.add(soundName);
